Add CloudEvents 1.0 specification-compliant DTOs for registry event webhook payloads. Includes base CloudEventDto with required fields (specversion, id, source, type, datacontenttype, data, time) and 8 wrapper classes for specific event types.
