+++ app/src/main/java/io/apicurio/registry/events/dto/ArtifactCreatedCloudEvent.java
@@ -6,6 +6,7 @@
 package io.apicurio.registry.events.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonUnwrapped;
 import io.quarkus.runtime.annotations.RegisterForReflection;
 
 import java.time.Instant;
@@ -17,22 +18,33 @@ import java.time.Instant;
 @RegisterForReflection
 public class ArtifactCreatedCloudEvent {
 
+    @JsonUnwrapped
     private CloudEventDto cloudEvent;
 
     public ArtifactCreatedCloudEvent() {
     }
 
     public static ArtifactCreatedCloudEvent from(io.apicurio.registry.events.ArtifactCreated event, String source) {
+        Instant eventTime = extractTimestampFromPayload(event.getPayload());
+        
         CloudEventDto dto = new CloudEventDto()
                 .withId(event.getId())
                 .withSource(source)
                 .withType("io.apicurio.registry.events.ArtifactCreated")
-                .withTime(Instant.now())
+                .withTime(eventTime)
                 .withData(event.getPayload());
 
         ArtifactCreatedCloudEvent wrapper = new ArtifactCreatedCloudEvent();
         wrapper.setCloudEvent(dto);
         return wrapper;
+    }
+
+    private static Instant extractTimestampFromPayload(org.json.JSONObject payload) {
+        if (payload != null && payload.has("createdOn")) {
+            long createdOn = payload.getLong("createdOn");
+            return Instant.ofEpochMilli(createdOn);
+        }
+        return Instant.now();
     }
 
     public CloudEventDto getCloudEvent() {
