+++ app/src/main/java/io/apicurio/registry/events/dto/CloudEventDto.java
