+++ app/src/test/java/io/apicurio/registry/events/dto/ArtifactDeletedCloudEventTest.java
