+++ app/src/test/java/io/apicurio/registry/events/dto/ArtifactMetadataUpdatedCloudEventTest.java
