+++ app/src/test/java/io/apicurio/registry/events/dto/ArtifactRuleConfiguredCloudEventTest.java
