+++ app/src/test/java/io/apicurio/registry/events/dto/ArtifactVersionStateChangedCloudEventTest.java
