+++ app/src/test/java/io/apicurio/registry/events/dto/GroupCreatedCloudEventTest.java
